breaking: rename folders 'permutation_sequences' and 'sequence-data'
