--- permutation_sequences/sequence-data/data-5-1.txt
@@ -1,8 +1,0 @@
-Number of sequences: 1
-Number of iterations: 4
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 1
-   Depth: 2, Node count: 1
-   Depth: 3, Node count: 1
-   Depth: 4, Node count: 1
