--- permutation_sequences/sequence-data/data-5-3.txt
@@ -1,15 +1,0 @@
-Number of sequences: 0
-Number of iterations: 31
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 0
-   Depth: 2, Node count: 0
-   Depth: 3, Node count: 1
-   Depth: 4, Node count: 2
-   Depth: 5, Node count: 4
-   Depth: 6, Node count: 7
-   Depth: 7, Node count: 9
-   Depth: 8, Node count: 8
-   Depth: 9, Node count: 0
-   Depth: 10, Node count: 0
-   Depth: 11, Node count: 0
