--- permutation_sequences/sequence-data/data-6-2.txt
@@ -1,19 +1,0 @@
-Number of sequences: 0
-Number of iterations: 5079
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 0
-   Depth: 2, Node count: 1
-   Depth: 3, Node count: 2
-   Depth: 4, Node count: 4
-   Depth: 5, Node count: 10
-   Depth: 6, Node count: 26
-   Depth: 7, Node count: 72
-   Depth: 8, Node count: 176
-   Depth: 9, Node count: 378
-   Depth: 10, Node count: 702
-   Depth: 11, Node count: 1152
-   Depth: 12, Node count: 1514
-   Depth: 13, Node count: 1042
-   Depth: 14, Node count: 0
-   Depth: 15, Node count: 0
