--- permutation_sequences/sequence-data/data-6-5.txt
@@ -1,13 +1,0 @@
-Number of sequences: 1
-Number of iterations: 5
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 0
-   Depth: 2, Node count: 0
-   Depth: 3, Node count: 0
-   Depth: 4, Node count: 0
-   Depth: 5, Node count: 1
-   Depth: 6, Node count: 1
-   Depth: 7, Node count: 1
-   Depth: 8, Node count: 1
-   Depth: 9, Node count: 1
