--- permutation_sequences/sequence-data/data-7-3.txt
@@ -1,40 +1,0 @@
-Number of sequences: 12255208
-Number of iterations: 2394316356
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 0
-   Depth: 2, Node count: 0
-   Depth: 3, Node count: 1
-   Depth: 4, Node count: 2
-   Depth: 5, Node count: 5
-   Depth: 6, Node count: 14
-   Depth: 7, Node count: 44
-   Depth: 8, Node count: 146
-   Depth: 9, Node count: 494
-   Depth: 10, Node count: 1661
-   Depth: 11, Node count: 5395
-   Depth: 12, Node count: 16837
-   Depth: 13, Node count: 50476
-   Depth: 14, Node count: 145454
-   Depth: 15, Node count: 401911
-   Depth: 16, Node count: 1061522
-   Depth: 17, Node count: 2656634
-   Depth: 18, Node count: 6300077
-   Depth: 19, Node count: 14215210
-   Depth: 20, Node count: 30641169
-   Depth: 21, Node count: 62404462
-   Depth: 22, Node count: 118243353
-   Depth: 23, Node count: 206657240
-   Depth: 24, Node count: 334605877
-   Depth: 25, Node count: 501991638
-   Depth: 26, Node count: 693936492
-   Depth: 27, Node count: 882568967
-   Depth: 28, Node count: 1040768201
-   Depth: 29, Node count: 1096924811
-   Depth: 30, Node count: 907367004
-   Depth: 31, Node count: 498429769
-   Depth: 32, Node count: 185977580
-   Depth: 33, Node count: 67145582
-   Depth: 34, Node count: 12255208
-   Depth: 35, Node count: 12255208
-   Depth: 36, Node count: 12255208
