--- permutation_sequences/sequence-data/data-7-5.txt
@@ -1,28 +1,0 @@
-Number of sequences: 0
-Number of iterations: 271
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 0
-   Depth: 2, Node count: 0
-   Depth: 3, Node count: 0
-   Depth: 4, Node count: 0
-   Depth: 5, Node count: 1
-   Depth: 6, Node count: 2
-   Depth: 7, Node count: 4
-   Depth: 8, Node count: 8
-   Depth: 9, Node count: 16
-   Depth: 10, Node count: 31
-   Depth: 11, Node count: 45
-   Depth: 12, Node count: 56
-   Depth: 13, Node count: 60
-   Depth: 14, Node count: 48
-   Depth: 15, Node count: 0
-   Depth: 16, Node count: 0
-   Depth: 17, Node count: 0
-   Depth: 18, Node count: 0
-   Depth: 19, Node count: 0
-   Depth: 20, Node count: 0
-   Depth: 21, Node count: 0
-   Depth: 22, Node count: 0
-   Depth: 23, Node count: 0
-   Depth: 24, Node count: 0
