--- permutation_sequences/sequence-data/data-8-5.txt
@@ -1,64 +1,0 @@
-Number of sequences: 1532076
-Number of iterations: 93239910      
-   // ADDENDUM: This is the number mod 2^32. I believe the actual number is 93239910 + 105 * 2^32 (451064805990)
-Depth Map:
-   Depth: 0, Node count: 0
-   Depth: 1, Node count: 0
-   Depth: 2, Node count: 0
-   Depth: 3, Node count: 0
-   Depth: 4, Node count: 0
-   Depth: 5, Node count: 1
-   Depth: 6, Node count: 2
-   Depth: 7, Node count: 5
-   Depth: 8, Node count: 14
-   Depth: 9, Node count: 41
-   Depth: 10, Node count: 121
-   Depth: 11, Node count: 346
-   Depth: 12, Node count: 933
-   Depth: 13, Node count: 2446
-   Depth: 14, Node count: 6287
-   Depth: 15, Node count: 15836
-   Depth: 16, Node count: 38854
-   Depth: 17, Node count: 96413
-   Depth: 18, Node count: 230474
-   Depth: 19, Node count: 539747
-   Depth: 20, Node count: 1241606
-   Depth: 21, Node count: 2777805
-   Depth: 22, Node count: 5929773
-   Depth: 23, Node count: 12408209
-   Depth: 24, Node count: 25297858
-   Depth: 25, Node count: 50588145
-   Depth: 26, Node count: 100132268
-   Depth: 27, Node count: 195262855
-   Depth: 28, Node count: 366350301
-   Depth: 29, Node count: 654404179
-   Depth: 30, Node count: 1110731914
-   Depth: 31, Node count: 1805293057
-   Depth: 32, Node count: 2883860124
-   Depth: 33, Node count: 297708654
-   Depth: 34, Node count: 2938971221
-   Depth: 35, Node count: 2452361713
-   Depth: 36, Node count: 2925717049
-   Depth: 37, Node count: 3357025507
-   Depth: 38, Node count: 3200295346
-   Depth: 39, Node count: 2713908487
-   Depth: 40, Node count: 2654148085
-   Depth: 41, Node count: 3816042783
-   Depth: 42, Node count: 1586630729
-   Depth: 43, Node count: 1525989574
-   Depth: 44, Node count: 2803759370
-   Depth: 45, Node count: 1590453420
-   Depth: 46, Node count: 2623140530
-   Depth: 47, Node count: 1966006607
-   Depth: 48, Node count: 1091269460
-   Depth: 49, Node count: 2978014657
-   Depth: 50, Node count: 1581154653
-   Depth: 51, Node count: 4046476003
-   Depth: 52, Node count: 1926130520
-   Depth: 53, Node count: 542931848
-   Depth: 54, Node count: 80163786
-   Depth: 55, Node count: 7469292
-   Depth: 56, Node count: 2239622
-   Depth: 57, Node count: 1532076
-   Depth: 58, Node count: 1532076
-   Depth: 59, Node count: 1532076
