--- permutation_sequences/sequences/sequence-1-1.csv
@@ -1,1 +1,0 @@
-0
