--- permutation_sequences/sequences/sequence-12-10.csv
@@ -1,2 +1,0 @@
-Number,of,sequences:,0
-
