--- permutation_sequences/sequences/sequence-13-11.csv
@@ -1,2 +1,0 @@
-Number,of,sequences:,0
-
