--- permutation_sequences/sequences/sequence-2-2.csv
@@ -1,1 +1,0 @@
-0,1
