--- permutation_sequences/sequences/sequence-3-3.csv
@@ -1,1 +1,0 @@
-0,1,2
