--- permutation_sequences/sequences/sequence-4-3.csv
@@ -1,1 +1,0 @@
-0,1,2,3,0,1
