--- permutation_sequences/sequences/sequence-5-2.csv
@@ -1,22 +1,0 @@
-0,1,2,3,4,2,0,4,1,3,0
-0,1,2,3,4,2,0,3,1,4,0
-0,1,2,3,4,1,3,0,4,2,0
-0,1,2,3,4,1,3,0,2,4,0
-0,1,2,3,4,0,3,1,4,2,0
-0,1,2,3,4,0,2,4,1,3,0
-0,1,2,3,1,4,3,0,4,2,0
-0,1,2,3,1,4,3,0,2,4,0
-0,1,2,3,1,4,2,0,4,3,0
-0,1,2,3,1,4,2,0,3,4,0
-0,1,2,3,1,4,0,3,4,2,0
-0,1,2,3,1,4,0,2,4,3,0
-0,1,2,3,0,4,3,1,4,2,0
-0,1,2,3,0,4,1,3,4,2,0
-0,1,2,3,0,2,4,3,1,4,0
-0,1,2,3,0,2,4,1,3,4,0
-0,1,2,0,3,4,2,3,1,4,0
-0,1,2,0,3,4,1,3,2,4,0
-0,1,2,0,3,2,4,3,1,4,0
-0,1,2,0,3,2,4,1,3,4,0
-0,1,2,0,3,1,4,3,2,4,0
-0,1,2,0,3,1,4,2,3,4,0
