--- permutation_sequences/sequences/sequence-5-4.csv
@@ -1,1 +1,0 @@
-0,1,2,3,4,0,1,2
