--- permutation_sequences/sequences/sequence-6-1.csv
@@ -1,1 +1,0 @@
-0,1,2,3,4,5
