+++ subset_sequences/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "permutation_sequences"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+num-integer = "0.1.46"
+rand = "0.8.5"
