+++ subset_sequences/sequence_data/data-4-2.txt
@@ -1,0 +1,10 @@
+Number of sequences: 0
+Number of iterations: 9
+Depth Map:
+   Depth: 0, Node count: 0
+   Depth: 1, Node count: 0
+   Depth: 2, Node count: 1
+   Depth: 3, Node count: 2
+   Depth: 4, Node count: 3
+   Depth: 5, Node count: 3
+   Depth: 6, Node count: 0
