+++ subset_sequences/sequence_data/data-5-2.txt
@@ -1,0 +1,14 @@
+Number of sequences: 22
+Number of iterations: 121
+Depth Map:
+   Depth: 0, Node count: 0
+   Depth: 1, Node count: 0
+   Depth: 2, Node count: 1
+   Depth: 3, Node count: 2
+   Depth: 4, Node count: 4
+   Depth: 5, Node count: 9
+   Depth: 6, Node count: 16
+   Depth: 7, Node count: 23
+   Depth: 8, Node count: 22
+   Depth: 9, Node count: 22
+   Depth: 10, Node count: 22
