+++ subset_sequences/sequence_data/data-6-4.txt
@@ -1,0 +1,21 @@
+Number of sequences: 0
+Number of iterations: 95
+Depth Map:
+   Depth: 0, Node count: 0
+   Depth: 1, Node count: 0
+   Depth: 2, Node count: 0
+   Depth: 3, Node count: 0
+   Depth: 4, Node count: 1
+   Depth: 5, Node count: 2
+   Depth: 6, Node count: 4
+   Depth: 7, Node count: 8
+   Depth: 8, Node count: 15
+   Depth: 9, Node count: 21
+   Depth: 10, Node count: 24
+   Depth: 11, Node count: 20
+   Depth: 12, Node count: 0
+   Depth: 13, Node count: 0
+   Depth: 14, Node count: 0
+   Depth: 15, Node count: 0
+   Depth: 16, Node count: 0
+   Depth: 17, Node count: 0
