+++ subset_sequences/sequence_data/data-7-2.txt
@@ -1,0 +1,25 @@
+Number of sequences: 541568
+Number of iterations: 3513131
+Depth Map:
+   Depth: 0, Node count: 0
+   Depth: 1, Node count: 0
+   Depth: 2, Node count: 1
+   Depth: 3, Node count: 2
+   Depth: 4, Node count: 4
+   Depth: 5, Node count: 10
+   Depth: 6, Node count: 27
+   Depth: 7, Node count: 87
+   Depth: 8, Node count: 291
+   Depth: 9, Node count: 975
+   Depth: 10, Node count: 3078
+   Depth: 11, Node count: 8994
+   Depth: 12, Node count: 24040
+   Depth: 13, Node count: 57794
+   Depth: 14, Node count: 122882
+   Depth: 15, Node count: 225474
+   Depth: 16, Node count: 356832
+   Depth: 17, Node count: 499200
+   Depth: 18, Node count: 588736
+   Depth: 19, Node count: 541568
+   Depth: 20, Node count: 541568
+   Depth: 21, Node count: 541568
