+++ subset_sequences/sequence_data/data-8-6.txt
@@ -1,0 +1,36 @@
+Number of sequences: 0
+Number of iterations: 735
+Depth Map:
+   Depth: 0, Node count: 0
+   Depth: 1, Node count: 0
+   Depth: 2, Node count: 0
+   Depth: 3, Node count: 0
+   Depth: 4, Node count: 0
+   Depth: 5, Node count: 0
+   Depth: 6, Node count: 1
+   Depth: 7, Node count: 2
+   Depth: 8, Node count: 4
+   Depth: 9, Node count: 8
+   Depth: 10, Node count: 16
+   Depth: 11, Node count: 32
+   Depth: 12, Node count: 63
+   Depth: 13, Node count: 93
+   Depth: 14, Node count: 120
+   Depth: 15, Node count: 140
+   Depth: 16, Node count: 144
+   Depth: 17, Node count: 112
+   Depth: 18, Node count: 0
+   Depth: 19, Node count: 0
+   Depth: 20, Node count: 0
+   Depth: 21, Node count: 0
+   Depth: 22, Node count: 0
+   Depth: 23, Node count: 0
+   Depth: 24, Node count: 0
+   Depth: 25, Node count: 0
+   Depth: 26, Node count: 0
+   Depth: 27, Node count: 0
+   Depth: 28, Node count: 0
+   Depth: 29, Node count: 0
+   Depth: 30, Node count: 0
+   Depth: 31, Node count: 0
+   Depth: 32, Node count: 0
